+++ models/src/v1/get.rs
@@ -10,7 +10,7 @@ pub fn tx_address<'a>(trits: &'a [Trit]) -> &'a [Trit] {
     &trits[ADDRESS_OFFSET..VALUE_OFFSET]
 }
 
-pub fn tx_value(trits: &[Trit]) -> isize {
+pub fn tx_value(trits: &[Trit]) -> i64 {
     num::trits2int(&trits[VALUE_OFFSET..TAG_OFFSET])
 }
 
@@ -18,16 +18,16 @@ pub fn tx_tag(trits: &[Trit]) -> &[Trit] {
     &trits[TAG_OFFSET..TIMESTAMP_OFFSET]
 }
 
-pub fn tx_timestamp(trits: &[Trit]) -> usize {
-    num::trits2int(&trits[TIMESTAMP_OFFSET..CURRENT_INDEX_OFFSET]) as usize
+pub fn tx_timestamp(trits: &[Trit]) -> u64 {
+    num::trits2int(&trits[TIMESTAMP_OFFSET..CURRENT_INDEX_OFFSET]) as u64
 }
 
-pub fn tx_current_index(trits: &[Trit]) -> usize {
-    num::trits2int(&trits[CURRENT_INDEX_OFFSET..LAST_INDEX_OFFSET]) as usize
+pub fn tx_current_index(trits: &[Trit]) -> u64 {
+    num::trits2int(&trits[CURRENT_INDEX_OFFSET..LAST_INDEX_OFFSET]) as u64
 }
 
-pub fn tx_last_index(trits: &[Trit]) -> usize {
-    num::trits2int(&trits[LAST_INDEX_OFFSET..BUNDLE_OFFSET]) as usize
+pub fn tx_last_index(trits: &[Trit]) -> u64 {
+    num::trits2int(&trits[LAST_INDEX_OFFSET..BUNDLE_OFFSET]) as u64
 }
 
 pub fn tx_bundle(trits: &[Trit]) -> &[Trit] {
